Wait for drain and ARR back to live bug fix
Two main changes in this commit. It seems there is a bug on IIS ARR 
where unavailable nodes come back to be available on their own after 
a while. 

To fix this, you need to setup a Health Check pointing to the main
entry point of the farm. This way the node pointing to the stopped
site in the farm will not serve requests even when ARR makes it 
available.

There other change is to wait 60 secs for the connection to drain
from the site about to be stopped.